app/requirements.txt
# Web Framework
Flask>=2.0.0
gunicorn>=20.1.0
python-decouple>=3.6

# AI/ML Libraries (commonly used - uncomment as needed)
# numpy>=1.24.0
# pandas>=2.0.0
# scikit-learn>=1.3.0
# tensorflow>=2.13.0
# torch>=2.0.0
# transformers>=4.30.0
# openai>=1.0.0
# langchain>=0.1.0

# Data Processing
# requests>=2.31.0
# python-dotenv>=1.0.0

